Add override flag to wipe_out command

+++ parsers/data.py
@@ -9,13 +9,21 @@ _, _, db = create_app()
 data = Manager(usage="Manage the data inside the database.")
 
 @data.command
-def wipe_out():
+def wipe_out(yes=False):
     '''
     Deletes and recreates the entire database using psql directly.
     Use it if 'data recreate' fails because the database needs to be migrated.
+    The command will prompt for confirmation unless the '--yes' option is provided
+    to override the prompt.
 
-    Recommended: run './manage data create' followed by './manage data populate' afterwards...
+    Example usage: 
+    ./manage data wipe_out --yes
+    ./manage data create
+    ./manage data populate
     '''
+    if not yes:
+        if not prompt_bool("Are you sure you want to lose all your data?"):
+            return
     database_name = 'rebase_web'
     check_call(['dropdb', database_name])
     check_call(['createdb', database_name])
@@ -24,7 +32,7 @@ def wipe_out():
 @data.command
 def drop():
     "Drops database tables"
-    if prompt_bool("Are you sure you want to lose all your data"):
+    if prompt_bool("Are you sure you want to lose all your data?"):
         db.drop_all()
